refactor the movielens examples
